Creates a couple of internal modules for the streaming
