+++ modules/internal/src/main/scala/higherkindness/mu/rpc/internal/client/package.scala
@@ -14,33 +14,31 @@
  * limitations under the License.
  */
 
-package higherkindness.mu.rpc
-package internal
+package higherkindness.mu.rpc.internal
 
-import io.grpc.{CallOptions, Channel, MethodDescriptor}
-import io.grpc.stub.ClientCalls
-import monix.execution.rstreams.Subscription
-import org.reactivestreams.{Publisher, Subscriber}
+import java.util.concurrent.{Executor => JavaExecutor}
+
+import cats.effect.Async
+import com.google.common.util.concurrent.{FutureCallback, Futures, ListenableFuture}
+
+import scala.concurrent.ExecutionContext
 
 package object client {
 
-  import higherkindness.mu.rpc.internal.converters._
+  private[internal] def listenableFuture2Async[F[_], A](
+      fa: => ListenableFuture[A])(implicit F: Async[F], EC: ExecutionContext): F[A] =
+    F.async { cb =>
+      Futures.addCallback(
+        fa,
+        new FutureCallback[A] {
+          override def onSuccess(result: A): Unit = cb(Right(result))
 
-  private[client] def createPublisher[Res, Req](
-      request: Req,
-      descriptor: MethodDescriptor[Req, Res],
-      channel: Channel,
-      options: CallOptions) = {
-    new Publisher[Res] {
-      override def subscribe(s: Subscriber[_ >: Res]): Unit = {
-        val subscriber: Subscriber[Res] = s.asInstanceOf[Subscriber[Res]]
-        s.onSubscribe(Subscription.empty)
-        ClientCalls.asyncServerStreamingCall(
-          channel.newCall[Req, Res](descriptor, options),
-          request,
-          subscriber.toStreamObserver
-        )
-      }
+          override def onFailure(t: Throwable): Unit = cb(Left(t))
+        },
+        new JavaExecutor {
+          override def execute(command: Runnable): Unit = EC.execute(command)
+        }
+      )
     }
-  }
+
 }
