+++ modules/internal/src/main/scala/higherkindness/mu/rpc/internal/server/unaryCalls.scala
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2017-2018 47 Degrees, LLC. <http://www.47deg.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package higherkindness.mu.rpc.internal.server
+import cats.effect.{Effect, IO}
+import io.grpc.stub.ServerCalls.UnaryMethod
+import io.grpc.stub.StreamObserver
+
+object unaryCalls {
+
+  def unaryMethod[F[_]: Effect, Req, Res](
+      f: Req => F[Res],
+      maybeCompression: Option[String]): UnaryMethod[Req, Res] =
+    new UnaryMethod[Req, Res] {
+      override def invoke(request: Req, responseObserver: StreamObserver[Res]): Unit = {
+        addCompression(responseObserver, maybeCompression)
+        Effect[F]
+          .runAsync(f(request))(either => IO(completeObserver(responseObserver)(either)))
+          .toIO
+          .unsafeRunAsync(_ => ())
+      }
+    }
+
+}
